mysql connection ???

+++ src/main/java/com/Models/JobformDAO.java
@@ -6,10 +6,11 @@ import java.util.List;
 
 public interface JobformDAO {
 
-   public void addJobform(JobForm jobForm);
+    public void addJobform(JobForm jobForm);
 
-   public List<JobForm> listJobform();
+    public List<JobForm> listJobform();
 
-   public void removeJobform(Integer id);
+    public void removeJobform(long id);
+
 
 }
